Add basic Tensorflow configuration
This adds Tensorflow in the following places:
  - Add dependency in pom.xml
  - Add helloTensorflow() test in Main
  - Update README instructions for new launch command

There's a big downside here: the jar file is no longer self-contained,
so we need to figure out a way to include tensorflow in the eventual
app. Right now, I added directions to launch the app via Maven, but we
probably don't want that for the long term.

Hopefully this works on Windows!