Add test helper and determinism tests
One of our criteria is that the same image outputs the same caption, so
I added a couple tests to verify this. This'll prevent us from breaking
this functionality as the captioning becomes more complex.

In the process, I added some image fixtures and a test image loader for
easy loading of test fixtures. I think this is cleaner than messing with
`File.open` in the actual test cases, and it makes the test purpose
clearer.