Add the Tensorflow model
This commit adds the trained (2M iter) tensorflow model to the
resources directory.

Because of their large size, the model files are managed by git lfs.
To download them:
  1. Download git-lfs (https://git-lfs.github.com/)
  2. Run `git lfs pull`
    2a. If that doesn't work, try deleting `src/main/resources/model`
        and trying again

If that doesn't work, I also added a shell script to download the
files: `bin/download_model.sh`. It's tested in OS X, and it should
work on UNIX-based systems that have curl and unzip. I'm not
sure about Windows there.

As a last-ditch option, you could go to:

  https://www.dropbox.com/s/1q7cwmewooi18s0/model.zip?dl=0

Unzip the files yourself, and put them in src/main/resources.
This will serve as a launching point for actually performing the
inference operation.

Trained Model Source:
  https://github.com/KranthiGV/Pretrained-Show-and-Tell-model